requirements.txt
Don't forget to download those packages:
django-ckeditor
fontawesomefree
cloudinary
cloudinary_storage
stripe